[TEMP COMMIT]. UI updates. Part 1. |NOT TESTED|

+++ composeApp/common/views/src/commonMain/kotlin/com/singing/app/common/views/shared/record/RecordCardActionsCallbacks.kt
@@ -9,12 +9,27 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.singing.app.common.views.model.state.RecordUiData
-import com.singing.app.common.views.views.generated.resources.*
+import com.singing.app.common.views.views.generated.resources.Res
+import com.singing.app.common.views.views.generated.resources.action_delete_record
+import com.singing.app.common.views.views.generated.resources.action_publish_record
+import com.singing.app.common.views.views.generated.resources.action_show_publication
+import com.singing.app.common.views.views.generated.resources.action_upload_record
+import com.singing.app.common.views.views.generated.resources.baseline_data_saver_on_24
+import com.singing.app.common.views.views.generated.resources.baseline_delete_outline_24
+import com.singing.app.common.views.views.generated.resources.baseline_file_upload_24
+import com.singing.app.common.views.views.generated.resources.baseline_open_in_new_24
 import org.jetbrains.compose.resources.stringResource
 import org.jetbrains.compose.resources.vectorResource
 
 
 private val chipHeight = Modifier.height(32.dp)
+
+data class RecordCardActionsState(
+    val canUpload: Boolean,
+    val canOpenPublication: Boolean,
+    val canPublish: Boolean,
+    val canDelete: Boolean,
+)
 
 data class RecordCardActionsCallbacks<T>(
     val onUploadRecord: (T) -> Unit,
@@ -26,9 +41,10 @@ data class RecordCardActionsCallbacks<T>(
 @Composable
 fun RecordCardActions(
     data: RecordUiData,
+    state: RecordCardActionsState?,
     actions: RecordCardActionsCallbacks<RecordUiData>,
 ) {
-    if (!data.isSavedRemote) {
+    if (!data.isSavedRemote && state?.canUpload == true) {
         AssistChip(
             modifier = chipHeight,
             label = {
@@ -51,7 +67,7 @@ fun RecordCardActions(
         )
     }
 
-    if (data.isPublished) {
+    if (data.isPublished && state?.canOpenPublication == true) {
         AssistChip(
             modifier = chipHeight,
             label = {
@@ -72,7 +88,9 @@ fun RecordCardActions(
                 actions.showPublication(data)
             }
         )
-    } else {
+    }
+
+    if (!data.isPublished && state?.canPublish == true) {
         AssistChip(
             modifier = chipHeight,
             label = {
@@ -95,24 +113,26 @@ fun RecordCardActions(
         )
     }
 
-    AssistChip(
-        modifier = chipHeight,
-        label = {
-            Text(
-                text = stringResource(Res.string.action_delete_record),
-                color = MaterialTheme.colorScheme.onSurface,
-                style = MaterialTheme.typography.labelLarge,
-            )
-        },
-        trailingIcon = {
-            Icon(
-                imageVector = vectorResource(Res.drawable.baseline_delete_outline_24),
-                tint = MaterialTheme.colorScheme.onSurface,
-                contentDescription = "",
-            )
-        },
-        onClick = {
-            actions.onDeleteRecord(data)
-        }
-    )
+    if (state?.canDelete == true) {
+        AssistChip(
+            modifier = chipHeight,
+            label = {
+                Text(
+                    text = stringResource(Res.string.action_delete_record),
+                    color = MaterialTheme.colorScheme.onSurface,
+                    style = MaterialTheme.typography.labelLarge,
+                )
+            },
+            trailingIcon = {
+                Icon(
+                    imageVector = vectorResource(Res.drawable.baseline_delete_outline_24),
+                    tint = MaterialTheme.colorScheme.onSurface,
+                    contentDescription = "",
+                )
+            },
+            onClick = {
+                actions.onDeleteRecord(data)
+            }
+        )
+    }
 }
